requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai 
langchain-community 
sentence-transformers
faiss-cpu
python-dotenv

